move pagination logic into useMemo
